Slice out message generalization factory.
