+++ lib/rtm-client.js
@@ -4,6 +4,7 @@ var PahoMQTT = require('../vendor/paho-mqtt');
 var merge = require('./utils/merge');
 var promised = require('./utils/promised');
 var events = require('./utils/events');
+var messageResponse = require('./responses/message');
 
 function RtmClient(authToken, options) {
   if (!(this instanceof RtmClient)) {
@@ -214,14 +215,14 @@ function _registerPahoClientEventListeners(rtmClient, pahoClient) {
   };
 
   pahoClient.onMessageDelivered = function(pahoMessage) {
-    var message = _getSdkMessage(pahoMessage);
+    var message = messageResponse.fromPahoMessage(pahoMessage);
 
     emit(message.channel + '/message-sent', message);
     emit('message-sent', message);
   };
 
   pahoClient.onMessageArrived = function(pahoMessage) {
-    var message = _getSdkMessage(pahoMessage);
+    var message = messageResponse.fromPahoMessage(pahoMessage);
 
     emit(message.channel + '/message', message);
     emit('message', message);
@@ -233,41 +234,6 @@ function _registerPahoClientEventListeners(rtmClient, pahoClient) {
   };
 
   return pahoClient;
-}
-
-function _getSdkMessage(pahoMessage) {
-  var data;
-  var decoded = false;
-  var error = null;
-
-  function decode() {
-    try {
-      data = JSON.parse(pahoMessage.payloadString);
-    } catch (err) {
-      error = err;
-    }
-  }
-
-  return {
-    channel: pahoMessage.topic,
-    raw: pahoMessage.payloadBytes,
-    get data() {
-      if (decoded === false) {
-        decode();
-        decoded = true;
-      }
-
-      return data;
-    },
-    get error() {
-      if (decoded === false) {
-        decode();
-        decoded = true;
-      }
-
-      return error;
-    }
-  };
 }
 
 module.exports = RtmClient;
